poe_set_port.go: Query live port configuration after set
Sometimes the user can specify a value that is out-of-bounds where
the switch will silently return SUCCESS but not commit the setting
(ex: due to it being out of bounds).

It might be possible to miss certain cases in the future, so query
the port configuration afterwards to report the actual state of the
switch ports.