port_set: Check that port name parameter is passed in before changing
Normally we pass all values back during POST, comparing and replacing
settings which are new. If the field has a length of 0, we assume
that the user has not set the value and send the old one.

However port name gets a pass on update, which allows a user to send a zero
length field to zero out the name, akin to entering the UI and deleting +
saving the field.

Unfortunately when updating other settings we also pass back all values,
but since the field length check is skipped for Port Name, we set it to
the passed in value or default (which is empty if the user did not define
it).

This is not the expected behavior and this commit checks that the flag is
set; if it is not, the old value is copied in to assume the expected behavior.